Back/auth

+++ backend/social-network-project/src/main/java/com/C9group34/socialnetworkproject/controllers/FavoritePublicationController.java
@@ -6,6 +6,7 @@ import com.C9group34.socialnetworkproject.exceptions.ExistingResourceException;
 import com.C9group34.socialnetworkproject.exceptions.ResourceNotFoundException;
 
 import com.C9group34.socialnetworkproject.service.FavoritePublicationService;
+import com.C9group34.socialnetworkproject.util.JWTutil;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -19,42 +20,57 @@ public class FavoritePublicationController {
 
     @Autowired
     private FavoritePublicationService favoritePService;
+    @Autowired
+    private JWTutil jwt;
 
-    @PostMapping("/{userId}/{publicationId}")
+    @PostMapping("/{publicationId}")
     @Transactional
-    public ResponseEntity create(@PathVariable Integer userId,
+    public ResponseEntity create(@RequestHeader(value = "Authorization") String token,
                                  @PathVariable Integer publicationId) {
 
         FavoritePublicationDto favoriteDto = null;
-        try {
-            favoriteDto = favoritePService.create( userId, publicationId);
-        } catch (ExistingResourceException e) {
-            System.out.println(e.getMessage());
-        } catch (ResourceNotFoundException e) {
-            System.out.println(e.getMessage());
+        String id = jwt.getKey(token);
+        if (jwt.verifyToken(token)) {
+            try {
+                favoriteDto = favoritePService.create(Integer.valueOf(id), publicationId);
+            } catch (ExistingResourceException e) {
+                System.out.println(e.getMessage());
+            } catch (ResourceNotFoundException e) {
+                System.out.println(e.getMessage());
+            }
+            return new ResponseEntity<>(favoriteDto.getId(), HttpStatus.CREATED);
         }
-        return new ResponseEntity<>(favoriteDto.getId(), HttpStatus.CREATED);
+        return new ResponseEntity("Acceso denegado", HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping
-    public ResponseEntity retrieve(@PathVariable Integer userId){
+    public ResponseEntity retrieve(@RequestHeader(value = "Authorization") String token) {
 
-        try {
-            return new ResponseEntity(favoritePService.retrieveAll(userId), HttpStatus.OK);
-        } catch (ResourceNotFoundException e) {
-            System.out.println(e.getMessage());
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        String id = jwt.getKey(token);
+        if (jwt.verifyToken(token)) {
+            try {
+                return new ResponseEntity(favoritePService.retrieveAll(Integer.valueOf(id)), HttpStatus.OK);
+            } catch (ResourceNotFoundException e) {
+                System.out.println(e.getMessage());
+                return new ResponseEntity(HttpStatus.NOT_FOUND);
+            }
         }
+
+        return new ResponseEntity("Acceso denegado", HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/{favoritePublicationId}")
-    public ResponseEntity delete(@PathVariable Integer favoritePublicationId){
-        try {
-            favoritePService.delete(favoritePublicationId);
-        } catch (ResourceNotFoundException e) {
-            throw new RuntimeException(e);
-        }
+    public ResponseEntity delete(@RequestHeader(value = "Authorization") String token,
+                                 @PathVariable Integer favoritePublicationId) {
 
-        return new ResponseEntity(HttpStatus.OK);
+        if (jwt.verifyToken(token)) {
+            try {
+                favoritePService.delete(favoritePublicationId);
+            } catch (ResourceNotFoundException e) {
+                throw new RuntimeException(e);
+            }
+            return new ResponseEntity(HttpStatus.OK);
+        }
+      return new ResponseEntity("Acceso denegado",HttpStatus.UNAUTHORIZED);
     }
 }
